requirements.txt
# Core Packages
flask==2.3.2
flask-socketio==5.3.6
python-dotenv==1.0.0

# Security
cryptography==42.0.5
pyopenssl==23.2.0

# Database
flask-sqlalchemy==3.0.5
python-dateutil==2.8.2

# Utilities
click==8.1.3
itsdangerous==2.1.2
jinja2==3.1.2
markupsafe==2.1.3
werkzeug==2.3.6

# Production Extras (uncomment when deploying)
# gunicorn==20.1.0
# eventlet==0.33.3
# redis==4.5.5